Fix timer-bug
When maxWait expires the current batch gets sent. Timer is then reset to maxWait and batchSize to zero. If no logs are sent during this maxWait, the timer expires but is not reset. After that any number of logs < config.BatchEntriesNumber is not sent, because there is not timer anymore.
Solution is to move maxWait.Reset out of the if-Block.